move logger to lib
